refactor user and sessions methods
